refactor(window): extract runtime submodule
Move runtime-heavy Window logic into window/runtime.rs so managed workspace lifecycle, endpoint event handling, and runtime connection presentation stop crowding the main window module.

Refs #98
